feat: create custom hook useShakeStatus.js
