+++ components/ShakeActivity.js
@@ -3,7 +3,8 @@ import { View, Text, ActivityIndicator } from "react-native";
 import styled from "styled-components/native"; // use /native when you are styling core components
 import { useFonts, Raleway_800ExtraBold } from "@expo-google-fonts/raleway";
 import { BORED_URL } from "../utils/Urls";
-import { Accelerometer } from "expo-sensors";
+// import { Accelerometer } from "expo-sensors";
+import { useShakeStatus } from "./useShakeStatus";
 
 // STYLED COMPONENTS
 
@@ -53,50 +54,64 @@ const Loader = styled.ActivityIndicator`
 `;
 
 const ShakeActivity = () => {
-  const [data, setData] = useState({
-    x: 0,
-    y: 0,
-    z: 0,
-  });
+  useShakeStatus();
   const [activity, setActivity] = useState({});
   const [loading, setLoading] = useState(false);
-  const [subscription, setSubscription] = useState(null);
   const [fontsLoaded] = useFonts({ Raleway_800ExtraBold });
 
   useEffect(() => {
     generateActivity();
   }, []);
 
-  useEffect(() => {
-    // This function determines how often our program reads the accelerometer data in milliseconds
-    Accelerometer.setUpdateInterval(1000);
-    // Start listening to the data when this SensorComponent is active
-    subscribe();
-    // Stop listening to the data when we leave SensorComponent
-    return () => unsubscribe();
-  }, []);
+  // useEffect(() => {
+  //   if (isShakingEnough(data)) {
+  //     generateActivity();
+  //   }
+  // }, [data]);
+  // const [data, setData] = useState({
+  //   x: 0,
+  //   y: 0,
+  //   z: 0,
+  // });
+  // const [activity, setActivity] = useState({});
+  // const [loading, setLoading] = useState(false);
+  // const [subscription, setSubscription] = useState(null);
+  // const [fontsLoaded] = useFonts({ Raleway_800ExtraBold });
 
-  useEffect(() => {
-    if (isShakingEnough(data)) {
-      generateActivity();
-    }
-  }, [data]);
+  // useEffect(() => {
+  //   generateActivity();
+  // }, []);
 
-  const subscribe = () => {
-    setSubscription(
-      Accelerometer.addListener((accelerometerData) => {
-        // Whenever this function is called, we have received new data
-        // The frequency of this function is controlled by setUpdateInterval
-        setData(accelerometerData);
-      })
-    );
-  };
+  // useEffect(() => {
+  //   // This function determines how often our program reads the accelerometer data in milliseconds
+  //   Accelerometer.setUpdateInterval(1000);
+  //   // Start listening to the data when this SensorComponent is active
+  //   subscribe();
+  //   // Stop listening to the data when we leave SensorComponent
+  //   return () => unsubscribe();
+  // }, []);
 
-  // This will tell the device to stop reading Accelerometer data.
-  const unsubscribe = () => {
-    subscription && subscription.remove();
-    setSubscription(null);
-  };
+  // useEffect(() => {
+  //   if (isShakingEnough(data)) {
+  //     generateActivity();
+  //   }
+  // }, [data]);
+
+  // const subscribe = () => {
+  //   setSubscription(
+  //     Accelerometer.addListener((accelerometerData) => {
+  //       // Whenever this function is called, we have received new data
+  //       // The frequency of this function is controlled by setUpdateInterval
+  //       setData(accelerometerData);
+  //     })
+  //   );
+  // };
+
+  // // This will tell the device to stop reading Accelerometer data.
+  // const unsubscribe = () => {
+  //   subscription && subscription.remove();
+  //   setSubscription(null);
+  // };
 
   const generateActivity = () => {
     setLoading(true);
@@ -106,24 +121,16 @@ const ShakeActivity = () => {
       .finally(() => setLoading(false));
   };
 
-  const isShakingEnough = (data) => {
-    const totalForce = Math.abs(data.x) + Math.abs(data.y) + Math.abs(data.z);
-    return totalForce > 1.78;
-  };
+  // const isShakingEnough = (data) => {
+  //   const totalForce = Math.abs(data.x) + Math.abs(data.y) + Math.abs(data.z);
+  //   return totalForce > 1.78;
+  // };
 
   if (loading || !fontsLoaded) {
     return <Loader size="large" color="#db7092" />;
   }
 
   return (
-    // <View>
-    //   <MainScreen
-    //     introText={"Lets..."}
-    //     titleText={activity.activity}
-    //     contentText={activity.activity}
-    //     shakeActivity={activity.type}
-    //   />
-    // </View>
     <Container style={{ fontFamily: "Raleway_800ExtraBold" }}>
       <PresentationText>Lets...</PresentationText>
       <TitleText>{activity.activity}</TitleText>
