+++ hooks/useShakeStatus.js
@@ -1,0 +1,43 @@
+import { useState, useEffect } from "react";
+import { Accelerometer } from "expo-sensors";
+
+// CUSTOMHOOK
+export const useShakeStatus = () => {
+  const [data, setData] = useState({
+    x: 0,
+    y: 0,
+    z: 0,
+  });
+
+  const [subscription, setSubscription] = useState(null);
+
+  useEffect(() => {
+    Accelerometer.setUpdateInterval(1000);
+    subscribe();
+    return () => unsubscribe();
+  }, []);
+
+  useEffect(() => {
+    if (isShakingEnough(data)) {
+      generateActivity();
+    }
+  }, [data]);
+
+  const subscribe = () => {
+    setSubscription(
+      Accelerometer.addListener((accelerometerData) => {
+        setData(accelerometerData);
+      })
+    );
+  };
+
+  const unsubscribe = () => {
+    subscription && subscription.remove();
+    setSubscription(null);
+  };
+
+  const isShakingEnough = (data) => {
+    const totalForce = Math.abs(data.x) + Math.abs(data.y) + Math.abs(data.z);
+    return totalForce > 1.78;
+  };
+};
